Now, when a position in the map is clicked, the character goes there.
Fixed Drawable types. Now the tiles in the map and the sprites inherit from the
same object. As they have now common values we can set a sprite to a tile's
position.

Improved level structure. It is much cleaner now, having a map of tiles. This
tiles contain both the tile drawable and the tile properties.

Removed a lot of the code from main. As an example now the function for checking
the mouse position is in the level, that contains the relevant information. The
function for updating the main character is also contained within. This also
implied removing lots of the global functions.